Restruct agent loader
Restruct the agent loader class so that the information of all agents is private, while the information of  agents that are being planned is public. This will be used for grouping agents.

The current version still plan paths for all agents at the same time. I wll later change CBS so that it can avoid collisions with already planned paths. Then we can use the grouping method.